[TRIVIAL] fix gcc compile warnings

./../build/../src/eventinterpreterplugins.cpp: In member function 'virtual void dss::EventInterpreterWebserviceGetWeatherInformationDone::result(long int, const std::string&)':
../../build/../src/eventinterpreterplugins.cpp:1050:12: warning: 'valOutdoorTemperature' may be used uninitialized in this function

../../build/../src/handler/system_states.cpp: In member function 'void dss::SystemState::callscene()':
../../build/../src/handler/system_states.cpp:498:51: warning: comparison between signed and unsigned integer expressions

../../build/../src/model/modulator.cpp: In member function 'void dss::DSMeter::publishToPropertyTree()':
../../build/../src/model/modulator.cpp:106:59: warning: dereferencing type-punned pointer will break strict-aliasing rules

See merge request !155

+++ src/handler/system_states.cpp
@@ -495,7 +495,7 @@ void SystemState::callscene() {
       // global or cluster-local states, but not both at the same time.
       // In the presence of cluster rain states inactive-rain on group0 serves
       // as a broadcast command only
-      for (size_t grp = GroupIDAppUserMin; grp <= GroupIDAppUserMax; grp++) {
+      for (int grp = GroupIDAppUserMin; grp <= GroupIDAppUserMax; grp++) {
         if (lookupState(state, formatAppartmentStateName(StateName::Rain, grp))) {
           state->setState(coSystem, State_Inactive);
         }
